Q12)lower,upper,digits,space,sym from sent
Consider that you are going to analyze the characters in the given string. Write 
a C program to extract the characters in the given string and print whether the 
character is an uppercase alphabet, lowercase alphabet, digits, whitespace, 
special symbols. Print the count of each category by storing their counts in an 
array. Use appropriate looping constructs to implement this.